validation and styling

+++ MovieLibraryAngular/ClientApp/src/app/movie-library/components/movie-library-details/movie-library-details.component.ts
@@ -6,22 +6,19 @@ import {ActivatedRoute, ParamMap, Router} from "@angular/router";
 @Component({
   selector: 'app-movie-details',
   templateUrl: 'movie-library-details.component.html',
-  providers: [ClientService]
+  providers: [ClientService],
+  styleUrls: ['movie-library-details.component.css']
 })
 
-export class MovieLibraryDetailsComponent implements OnInit{
-  constructor(private client: ClientService, private route: ActivatedRoute, private router: Router) {
-  }
-  public movie: MovieDetails;
-  private id: number;
+export class MovieLibraryDetailsComponent implements OnInit {
+
+  public movie: MovieDetails = new MovieDetails();
+  public id: number;
   public actors: string;
 
-  getMovie(): MovieDetails{
-    return this.route.snapshot.data['movie'];
-  }
+  public readonly defaultImageUrl = 'https://media.istockphoto.com/id/1216251206/vector/no-image-available-icon.jpg?b=1&s=170667a&w=0&k=20&c=MiWLEcUdxZONMlnsN_k5OCaz_nLviJB0Hvcz5Wlp5oI=';
 
-  getActors(): string{
-     return this.movie.actors.map(actor => actor.name + ' ' + actor.surname).join(', ');
+  constructor(private client: ClientService, private route: ActivatedRoute, private router: Router) {
   }
 
   ngOnInit(): void {
@@ -30,9 +27,19 @@ export class MovieLibraryDetailsComponent implements OnInit{
     this.actors = this.getActors()
   }
 
-  deleteMovie(){
-    this.client.deleteMovie(this.id).subscribe(() => {
-      this.router.navigate(['/movie']);
-    })
+  getMovie(): MovieDetails {
+    return this.route.snapshot.data['movie'];
+  }
+
+  getActors(): string {
+    return this.movie.actors.map(actor => actor.name + ' ' + actor.surname).join(', ');
+  }
+
+  deleteMovie() {
+    if (confirm("Are you sure to delete this movie?")) {
+      this.client.deleteMovie(this.id).subscribe(() => {
+        this.router.navigate(['/movie']);
+      })
+    }
   }
 }
